+++ pkg/apis/deployment/v2alpha1/zz_generated.deepcopy.go
@@ -2710,6 +2710,11 @@ func (in *ServerGroupSpec) DeepCopyInto(out *ServerGroupSpec) {
 		*out = new(ServerGroupUpgradeMode)
 		**out = **in
 	}
+	if in.RestartPolicy != nil {
+		in, out := &in.RestartPolicy, &out.RestartPolicy
+		*out = new(v1.RestartPolicy)
+		**out = **in
+	}
 	return
 }